Avoid Cyclic dependency
